requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

braindecode==1.3.2
h5py==3.16.0
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
torch==2.11.0
umap-learn==0.5.11
